requirements-test.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
